Implemented error rendering
