+++ index.html
@@ -42,7 +42,7 @@
             <span class="loader"></span>
             <div class="task__group" id="task__container" data-task_lists></div>
 
-            <div class="task__actions">
+            <div class="task__actions" id="task__actions">
                 <span data--indicator>...</span>
 
                 <div class="links">
