+++ lib/index.js
@@ -2,16 +2,16 @@ import '../scss/style.scss';
 import { listenForTodosConnectionStatus, 
         listenForTodosConnectionFailed
  } from './listeners';
-import { renderAuthModal, renderTodos, resetAll } from './renderer';
+import { renderAuthModal, renderTodos, signOut } from './renderer';
 import { 
     createTodo,
     registerTaskConnectionListener, registerTasksUpdateListener,
-    registerTaskFailedConnectionListener, handleSignOut 
+    registerTaskFailedConnectionListener 
 } from './script';
 
 import('./drag_n_drop');
 
-const signOut = ()=> handleSignOut(resetAll);
+
 
 // Run when the browser is done loading
 window.onload = (e) => {
