+++ lib/listeners.js
@@ -2,6 +2,7 @@
 
 // Authentication Listeners
 
+import { renderTodosError } from "./renderer";
 import { authenticate,createTaskSocket, USER_TOKEN } from "./script";
 
 export function listenForAuthSwitch (){
@@ -128,7 +129,10 @@ export const listenForTodosConnectionStatus = (connecting=true)=>{
 }
 
 export const listenForTodosConnectionFailed = (error, signOutHandler)=>{
-    if (error === 'auth_error') return signOutHandler();
+    if (error === 'auth_error') {
+        renderTodosError("Authentication is required!");
+        return signOutHandler();
+    }
 };
