+++ lib/renderer.js
@@ -2,12 +2,15 @@
 
 import {
     clearCompletedTodo,
-    deleteTodo, updateTodo, USER_TOKEN
+    deleteTodo, handleSignOut, updateTodo, USER_TOKEN
 } from './script';
-import { alreadySignedInTemplate, signUpTemplate, signInTemplate,renderTodosTemplate, initiateCloseModal, taskTemplate } from './templates';
+import { alreadySignedInTemplate, signUpTemplate, signInTemplate, initiateCloseModal, taskTemplate, taskLoadingTemplate } from './templates';
 import { listenForAuthSwitch, listenForSignin, 
-        listenForSignOut, listenForSignup, 
+        listenForSignOut, listenForSignup,
  } from './listeners';
+
+export const signOut = ()=> handleSignOut(resetAll);
+
 // ================== Tasks Actions =================================
 // Filters
 const undone_filter = document.querySelector('[data-filter--undone]');
@@ -72,14 +75,49 @@ clear_completed.addEventListener('click', (e) => {
 });
 
 // ==================================================================
+// Render loading template
+export const renderTodosTemplate = (amount=4) => {
 
+    const todoListContainer = document.querySelector('[data-task_lists]');
+    let content = '';
+
+    for (let i = 0; i <= amount; i++) content += taskLoadingTemplate;
+
+    todoListContainer.innerHTML = content;
+
+    document.querySelector("[data--indicator]").innerHTML = `loading tasks...`;
+    // document.querySelector(".task-items").setAttribute('data-loading', 'true');
+
+}
+
+// Render loading template
+export const renderTodosError = (message) => {
+
+    const todoListContainer = document.querySelector('[data-task_lists]');
+    let content = `
+
+        <div class="task-error">
+
+            <p>${message}</p>
+        </div>
+
+    `
+
+    todoListContainer.innerHTML = content;
+
+    document.querySelector("#task__actions").style.display = "none";
+
+}
 
 // Render all todos in DOM
 // Receives all the todos to render
-export const renderTodos = (incoming_tasks = []) => {
+export const renderTodos = (incoming_tasks = null) => {
+
     todos = incoming_tasks;
 
-    if (todos.length < 1) return renderTodosTemplate();
+    if (!Boolean(todos)) return renderTodosTemplate();
+
+    if (todos.length < 1) return renderTodosError("Your created task will show here");
 
     const todoListContainer = document.querySelector('[data-task_lists]');
     let content = '';
@@ -96,6 +134,7 @@ export const renderTodos = (incoming_tasks = []) => {
     todoListContainer.innerHTML = content;
     trackTodos();
 
+    document.querySelector("#task__actions").style.display = "flex";
     document.querySelector("[data--indicator]").innerHTML = `${filteredTodo.length} item${filteredTodo.length > 1 ? 's':''} left`;
 
 }
