+++ lib/templates.js
@@ -143,51 +143,13 @@ export const signUpTemplate = `
     </div>
 `
 
-// Render loading template
-export const renderTodosTemplate = () => {
+export const taskLoadingTemplate = `
+    <div class="task loading">
+        <span class="checker"></span>
 
-    const todoListContainer = document.querySelector('[data-task_lists]');
-    let content = '';
-
-    for (let i = 0; i <= 4; i++) {
-        content += `
-
-        <div class="task loading">
-            <span class="checker"></span>
-
-            <p></p>
-        </div>
-
-        `
-    };
-
-    todoListContainer.innerHTML = content;
-
-    document.querySelector("[data--indicator]").innerHTML = `loading tasks...`;
-    // document.querySelector(".task-items").setAttribute('data-loading', 'true');
-
-}
-
-// Render loading template
-export const renderTodosError = (message) => {
-
-    const todoListContainer = document.querySelector('[data-task_lists]');
-    let content = `
-
-        <div class="task-error">
-
-            <p>Could not load your tasks</p>
-
-            <button> Try again </button>
-        </div>
-
-    `
-
-    todoListContainer.innerHTML = content;
-
-    document.querySelector("[data--indicator]").innerHTML = `...`;
-
-}
+        <p></p>
+    </div>
+`
 
 // Take a event watch over close modal when it is rendered to DOM
 export const initiateCloseModal = (modal) =>{
